requirements.txt
# ATLAS Security Assessment Framework
# Core Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# CLI
typer[all]>=0.9.0
rich>=13.0.0

# Reconnaissance
python-nmap>=0.7.1

# HTTP Requests
requests>=2.31.0
httpx>=0.25.0

# Database
aiosqlite>=0.19.0

# Templating & Reporting
jinja2>=3.1.0

# Utilities
pydantic>=2.5.0
python-multipart>=0.0.6

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
